+++ src/compute-engine/latex-syntax/dictionary/definitions-sets.ts
@@ -103,12 +103,12 @@ export const DEFINITIONS_SETS: LatexDictionary = [
       // accept arguments that are `Set`
       const ce = parser.computeEngine!;
 
-      if (!ce || !ce.box(lhs).domain.isCompatible('Sets')) return null;
+      if (!ce || !ce.box(lhs).domain?.isCompatible('Sets')) return null;
 
       const index = parser.index;
       const rhs = parser.parseExpression({ ...until, minPrec: 390 });
       // If the rhs argument is not a set, bail
-      if (rhs === null || ce.box(lhs).domain.isCompatible('Sets') !== true) {
+      if (rhs === null || ce.box(lhs).domain?.isCompatible('Sets') !== true) {
         parser.index = index;
         return null;
       }
@@ -159,7 +159,6 @@ export const DEFINITIONS_SETS: LatexDictionary = [
     kind: 'matchfix',
     openTrigger: '{',
     closeTrigger: '}',
-    precedence: 20,
     // @todo: the set syntax can also include conditions...
     parse: (_parser: Parser, body: Expression): Expression => {
       if (body === null || isEmptySequence(body)) return 'EmptySet';