+++ src/compute-engine/library/logic.ts
@@ -95,7 +95,10 @@ export const LOGIC_LIBRARY: IdentifierDefinitions = {
       evaluate: processImplies,
     },
   },
-  Exists: { signature: { domain: 'Functions' } },
+  Exists: { signature: { domain: 'Functions' }, hold: 'all' },
+  ExistsUnique: { signature: { domain: 'Functions' }, hold: 'all' },
+  ForAll: { signature: { domain: 'Functions' }, hold: 'all' },
+  ForAllUnique: { signature: { domain: 'Functions' }, hold: 'all' },
 };
 
 function processAnd(